+++ src/runtime/ink_story_state.cpp
@@ -6,8 +6,8 @@
 
 InkObject* InkStoryState::get_current_object(std::int64_t index_offset) {
 	std::int64_t index = index_in_knot() + index_offset;
-	if (index >= 0 && index < current_knot_size()) {
-		return current_knot().knot->objects[index];
+	if (index >= 0 && index < static_cast<std::int64_t>(current_knot_size())) {
+		return current_knot().knot->objects[static_cast<std::size_t>(index)];
 	}
 
 	return nullptr;
@@ -31,6 +31,6 @@ InkStoryState::KnotStatus& InkStoryState::current_nonchoice_knot() {
 }
 
 void InkStoryState::increment_visit_count(const std::string& knot) {
-	std::size_t current_count = variables[knot].asInt();
+	std::size_t current_count = static_cast<std::size_t>(variables[knot].asInt());
 	variables[knot] = current_count + 1;
 }
